devtools/server/tests/mochitest/test_styles-modify.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=
-->
<head>
  <meta charset="utf-8">
  <title>Test for Bug </title>

  <script type="application/javascript" src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="chrome://mochikit/content/tests/SimpleTest/test.css">
  <script type="application/javascript" src="inspector-helpers.js"></script>
  <script type="application/javascript">
"use strict";

const {isCssPropertyKnown} = require("devtools/server/actors/css-properties");

window.onload = function() {
  SimpleTest.waitForExplicitFinish();
  runNextTest();
};

var gWalker = null;
var gStyles = null;
var gInspectee = null;

addAsyncTest(async function setup() {
  const url = document.getElementById("inspectorContent").href;
  let inspector;

  await new Promise(resolve => {
    attachURL(url, function(err, client, tab, doc) {
      gInspectee = doc;
      const {InspectorFront} = require("devtools/shared/fronts/inspector");
      inspector = InspectorFront(client, tab);
      resolve();
    });
  });

  gWalker = await inspector.getWalker();
  gStyles = await inspector.getPageStyle();

  runNextTest();
});

addAsyncTest(async function modifyProperties() {
  const localNode = gInspectee.querySelector("#inheritable-rule-inheritable-style");

  const node = await gWalker.querySelector(gWalker.rootNode,
    "#inheritable-rule-inheritable-style");

  const applied = await gStyles.getApplied(node,
    { inherited: false, filter: "user" });

  const elementStyle = applied[0].rule;
  is(elementStyle.cssText, localNode.style.cssText, "Got expected css text");

  // Change an existing property...
  await setProperty(elementStyle, 0, "color", "black");
  // Create a new property
  await setProperty(elementStyle, 1, "background-color", "green");

  // Create a new property and then change it immediately.
  await setProperty(elementStyle, 2, "border", "1px solid black");
  await setProperty(elementStyle, 2, "border", "2px solid black");

  is(elementStyle.cssText,
     "color: black; background-color: green; border: 2px solid black;",
     "Should have expected cssText");
  is(elementStyle.cssText, localNode.style.cssText,
     "Local node and style front match.");

  // Remove all the properties
  await removeProperty(elementStyle, 0, "color");
  await removeProperty(elementStyle, 0, "background-color");
  await removeProperty(elementStyle, 0, "border");

  is(elementStyle.cssText, "", "Should have expected cssText");
  is(elementStyle.cssText, localNode.style.cssText,
     "Local node and style front match.");

  runNextTest();
});

async function setProperty(rule, index, name, value) {
  const changes = rule.startModifyingProperties(isCssPropertyKnown);
  changes.setProperty(index, name, value);
  await changes.apply();
}

async function removeProperty(rule, index, name) {
  const changes = rule.startModifyingProperties(isCssPropertyKnown);
  changes.removeProperty(index, name);
  await changes.apply();
}

addTest(function cleanup() {
  gStyles = null;
  gWalker = null;
  gInspectee = null;
  runNextTest();
});

  </script>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=">Mozilla Bug </a>
<a id="inspectorContent" target="_blank" href="inspector-styles-data.html">Test Document</a>
<p id="display"></p>
<div id="content" style="display: none">

</div>
<pre id="test">
</pre>
</body>
</html>
